+++ kitty/config.py
@@ -293,7 +293,6 @@ type_map = {
     'macos_option_as_alt': to_bool,
     'macos_titlebar_color': macos_titlebar_color,
     'box_drawing_scale': box_drawing_scale,
-    'x11_bell_volume': int,
     'background_opacity': unit_float,
     'tab_separator': tab_separator,
     'active_tab_font_style': tab_font_style,